revise create and delete from S3 in service

+++ webapp/NoteTaking/src/main/java/edu/neu/coe/csye6225/service/AmazonS3ClientServiceImpl.java
@@ -7,6 +7,8 @@ import com.amazonaws.auth.profile.ProfileCredentialsProvider;
 import com.amazonaws.services.s3.model.*;
 
 import edu.neu.coe.csye6225.entity.Attachment;
+import edu.neu.coe.csye6225.entity.Note;
+import edu.neu.coe.csye6225.mapper.AttachmentMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,12 +26,15 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.net.URL;
 
 @Component
 public class AmazonS3ClientServiceImpl implements AmazonS3ClientService{
     private String awsS3AudioBucket;
     private AmazonS3 amazonS3;
     private static final Logger logger = LoggerFactory.getLogger(AmazonS3ClientServiceImpl.class);
+    @Autowired
+    private AttachmentMapper attachmentMapper;
 
 //    @Autowired
 //    public AmazonS3ClientServiceImpl(Region awsRegion,  AWSCredentialsProvider awsCredentialsProvider, String awsS3AudioBucket)
@@ -61,16 +66,15 @@ public class AmazonS3ClientServiceImpl implements AmazonS3ClientService{
     }
 
     @Async
-    public void createAttachmentToS3Bucket(MultipartFile multipartFile, boolean enablePublicReadAccess)
+    public boolean createAttachmentToS3Bucket(Note note, MultipartFile multipartFile, boolean enablePublicReadAccess)
     {
         long fileSize = multipartFile.getSize();
         String fileName = multipartFile.getOriginalFilename();
+        String noteId = note.getNoteId();
 
         try {
             //creating the file in the server (temporarily)
             File file = new File(fileName);
-//            if(!file.exists())
-//                return false;
             String fileType = fileName.substring(fileName.lastIndexOf("."),fileName.length());
             FileOutputStream fos = new FileOutputStream(file);
             fos.write(multipartFile.getBytes());
@@ -78,32 +82,42 @@ public class AmazonS3ClientServiceImpl implements AmazonS3ClientService{
 
             PutObjectRequest putObjectRequest = new PutObjectRequest(this.awsS3AudioBucket, fileName, file);
 
-//            if (enablePublicReadAccess) {
-//                putObjectRequest.withCannedAcl(CannedAccessControlList.PublicRead);
-//            }
             PutObjectResult result = this.amazonS3.putObject(putObjectRequest);
+            URL url = amazonS3.getUrl(awsS3AudioBucket,fileName);
             String eTag = result.getETag();
             if(StringUtils.isNotBlank(eTag))
-                logger.info("Send Attachment to Amazon S3 succeeded, ETag is "+result.getETag());
-            // TODO
-            // CALL createAttachment to store file info in db
-//            Attachment attachment = new Attachment()
+                logger.info("Send Attachment to Amazon S3 succeeded, Etag is "+eTag);
+            Attachment attachment = new Attachment(noteId,url.toString(),fileSize,fileType,fileName,eTag);
+            if (attachmentMapper.insertAttachment(attachment)<=0){
+                logger.error("meta data failed to insert into database!");
+                return false;
+            }
+            logger.info("Attachment information has been successfully saved to database");
             //removing the file created in the server
             file.delete();
-//            return true;
+            return true;
         } catch (IOException | AmazonServiceException ex) {
             logger.error("error [" + ex.getMessage() + "] occurred while uploading [" + fileName + "] ");
-//            return false;
+            return false;
         }
     }
 
     @Async
-    public void deleteAttachmentFromS3Bucket(String fileName)
+    public boolean deleteAttachmentFromS3Bucket(Note note, String attachmentId)
     {
+        Attachment attachment = attachmentMapper.getAttachmentById(attachmentId);
         try {
-            amazonS3.deleteObject(new DeleteObjectRequest(awsS3AudioBucket, fileName));
+            amazonS3.deleteObject(new DeleteObjectRequest(awsS3AudioBucket, attachment.getFileName()));
+            logger.info("delete attachment from S3 succeeded");
+            if(attachmentMapper.deleteAttachment(attachment)<=0){
+                logger.error("delete attachment from database failed");
+                return false;
+            }
+            logger.info("delete attachment from database succeeded");
+            return true;
         } catch (AmazonServiceException ex) {
-            logger.error("error [" + ex.getMessage() + "] occurred while removing [" + fileName + "] ");
+            logger.error("error [" + ex.getMessage() + "] occurred while removing [" + attachment.getFileName() + "] ");
+            return false;
         }
     }
 }
